refactor: improve code separation in details page
