--- src/entities/Product.ts
@@ -1,19 +1,0 @@
-export interface Product {
-    id: number;
-    image: string;
-    hoverImage: string;
-    title: string;
-    brand: string;
-    discount: string;
-    originalPrice: string;
-    price: string;
-    seller: string;
-    rating: string;
-    description: string;
-    technicalSpecifications: Specification[];
-  }
-  
-  export interface Specification {
-    feature: string;
-    details: string[];
-  }
