+++ src/pages/ProductDescriptionPage/ProductDescription.tsx
@@ -5,15 +5,15 @@ import { TechnicalSpecifications } from "./TechnicalSpecifications";
 import DynamicBreadcrumb from "@/components/ui/dynamicbreadcrumb";
 import { ProductRecommendations } from "./ProductRecommendations";
 
-
 export function ProductDescription() {
-    const { id } = useParams();
-    return (
-        <MainLayout>
-            <DynamicBreadcrumb />
-            <ProductDetail id={Number(id)} />
-            <TechnicalSpecifications id={Number(id)} />
-            <ProductRecommendations currentProductId={Number(id)} />
-        </MainLayout>
-    );
+  const { id } = useParams();
+
+  return (
+    <MainLayout>
+      <DynamicBreadcrumb />
+      <ProductDetail id={Number(id)} />
+      <TechnicalSpecifications id={Number(id)} />
+      <ProductRecommendations currentProductId={Number(id)} />
+    </MainLayout>
+  );
 }
